runmodes: let thread count callback return uint16_t
It will be a long time before we need more than 64k threads.

Update capture methods.

+++ src/source-af-xdp.h
@@ -29,7 +29,7 @@
 typedef struct AFXDPIfaceConfig {
     char iface[AFXDP_IFACE_NAME_LENGTH];
     /* number of threads */
-    int threads;
+    uint16_t threads;
     int promisc;
 
     /* misc use flags */
